soc: qcom: bam_dmux: Correct payload length handing for dynamic mtu
Packets received under the dynamic mtu feature do not have a valid payload
size in the header, so the correct size is computed based on data from
SPS. This computation is incorrect, which results in the client
processing extra data in the scenario where their protocol requires the
packet to be dropped.

Correct the payload length handling for this scenario.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Arun Kumar Neelakantam <aneela@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: FireLord1 <firelord.xda@gmail.com>